Implementación de funciones nativas y mejoras en el análisis semántico
- Se añadieron funciones nativas como Abs, Min, Max, Potencia, entre otras, con sus respectivos tipos de retorno y argumentos.
- Se mejoró la verificación de tipos en llamadas a funciones nativas en el análisis semántico.
- Se implementó la búsqueda de tipos personalizados en el árbol AST.
- Se corrigió el manejo de constantes enteras para soportar valores de 64 bits.
- Se añadió un nuevo archivo de runtime para manejar solicitudes HTTP.
- Se implementaron pruebas para las nuevas funciones nativas y tipos personalizados.
- Se mejoró la impresión de valores constantes en la jerarquía de tablas de símbolos.
- Se implementaron nuevas pruebas para arrays, funciones, y procedimientos void.

+++ src/include/symbols.h
@@ -4,6 +4,7 @@
 #define CUBETAS_TABLA_SIMBOLOS 100
 
 #include "types.h"
+#include <stdint.h>
 
 typedef struct EntradaSimbolo
 {
@@ -12,7 +13,7 @@ typedef struct EntradaSimbolo
     int es_constante;
     union
     {
-        int valor_int;
+        int64_t valor_int;
         double valor_float;
         char *valor_cadena;
         int valor_bool;
